moves unused assigner features into train_assigner
